Gate-Fix: Replace Octavia hacking O349 with H216
The new flake8 includes check H216 to encourage the use of unittest.mock
instead of third party mock modules. Octavia had a hacking check, O349,
for the same purpose. Unfortunately, H216 is flagging that check as
an improper use.
This patch removes O349, in favor of H216, for the Octaiva hacking checks.

Change-Id: I5c11c892ca1184624d6d8e111f69ec30cab2a0d8