Optimize CountPoolChildrenForQuota task in amphorav2
The CountPoolChildrenForQuota task takes 3 or 4 seconds to execute in
amphorav2. When cascade deleting a load balancer with many pools (60),
it might take minutes.

We don't need to get and to convert all the fields from the object to
compute the number of children of a pool, we can optimize this task by
creating a new method in the PoolRepository class.

Story 2008873
Task 42414

Change-Id: Ie4d40509ae7e6faccecd1d0cc0bc4c7d93959261