Fix race conditions between API and worker DB calls
The Octavia API creates new resource in the DB and passes the resource
and its ID to the worker. The worker then accesses to the resource using
the database, but it might try to fetch the element before it's
commited by the API.
This commit ensures that the resource has been commited to the DB before
processing the request.

Story 2008878
Task 42423

Change-Id: I55136f570db33cfd22e919019c18e9028a6a9bf7
(cherry picked from commit cd272ae5a948c22e26b0ea678ea452f559e14d62)