Fix invalid/deleted members in haproxy config file
Fix the member list in the haproxy configuration files, it prevents
DELETED members to be in the list.

This patch is specific to the stable/train branch. The fix for other
releases was introduced in stable/ussuri but it is part of a patch
(Ic44019b8877f008e6d7a75ceed1b7fd958e051d0) that is not backportable to
stable/train. Only the diffs related to the jinja templates have been
included in this new commit.

One part of the fix was already backported to stable/train in
Ib7b083e1dfbfd7afcca870ed6f60a871b2e19253

Story 2008871
Task 42402

Change-Id: I4c85425a774594c52a0bb743fd5b787706201425