Workaround for functional tests timeouts in DriverAgentTest
Some DriverAgentTest tests have been failing in CI with a timeout:

      File /home/zuul/src/opendev.org/openstack/octavia/.tox/functional-py37/\
        lib/python3.7/site-packages/octavia_lib/api/drivers/driver_lib.py, \
        line 61, in _recv
    char = sock.recv(1)

    socket.timeout: timed out~

It appears that, on CI servers, the SQLAlchemy queries are really slow.
The tests are launched in parallel which makes the things worse. Queries
that take less than 200ms on a dev environment might hit the socket
timeout (5 seconds).

This patch adds a workaround for the functional tests, it mocks the
default (and hardcoded) socket timeout and sets it to 30sec.

Change-Id: If048bc0eba34dc227cbd17561a61f43665f7b10e