Fix duplicate SG creation for listener peer port
In cases where the listener protcol port is same as the peer port
and allowed_cidr set to 0.0.0.0/0 explicitly, the listener is not
provisioned due to duplicate security group creation for peer port
with None as remote_ip_prefix. Neutron SG defaults remote_ip_prefix
to 0.0.0.0/0 if not specified or None and hence the error SG rule
already exists.

Remove the duplicate entry from the updated_ports.

Story: #2009117
Change-Id: I9dbdb71e9b94bbcc75766a8687a996d5358f3381