Preserve haproxy server states during reloads
haproxy doesn't keep the state of the servers (UP or DOWN) during
reloads, so a member with an ERROR operating_status may be updated
to ONLINE after updating a load balancer.

This commit adds the load-server-state-from-file option in haproxy
configuration files. It also adds an extra step in the haproxy systemd
service file to dump the current server-state in a file when reloading
the service.

Story: 2009142
Task: 43084

Change-Id: Ia8ec48ab858eeecf71ed429e5b7625fd7af9d8f6
(cherry picked from commit 1b15e1ef22963324e556e072f4bc58021f96bfec)