Fix PlugVIPAmphora revert function in amphorav2
This function should call the unplug_aap_port function that takes
DB models as arguments, but the subnet is not converted and is passed as
a provider dict. It triggers a "'dict' object has no attribute
'network_id'" error when reverting a load balancer creation.
Ensure that the subnet is converted from dict to DB model.

Story 2009157
Task 43132

Change-Id: I4bec17a484d0d39c9cca389421c63d378c36353c