Fix template issue with user_data_config_drive
Fix a python3 error that prevented to use the user_data_config_drive
option when building amphorae.

user_data_config_drive has been broken since we switched to python3,
a mix of str and bytes variables were provided to the jinja2 templates.
This commit ensures that the type of all the certificates provided to
jinja2 are strings.

Story: 2005553
Task: 30705

Change-Id: I53ed0425f00cfbbc964f5083b4f2942cf5bf565e