Enable taskflow retry feature when waiting for compute
The ComputeActiveWait task may be blocked for 300 seconds (with the
default settings for amp_active_retries and amp_active_wait_sec) but
that triggers trigger issues when persistence is enabled if this delay
exceeds the jobboard_expiration_time value.

ComputeActiveWait is renamed to ComputeWait and now uses the retry
feature of taskflow so it would not block taskflow for a long period.

Story 2008956
Task 42584

Change-Id: I174e63cfe176e1792d0bbcd348dbc9115544e297