Remove incorrect info message
In amphorav2 the controller now waits for the completion of the job to
send a keepalive to taskflow.  This patch is a follow up patch to
I6619ba117f7051fe44086389789bc6e2810fd23d, it removes a INFO message
that no longer makes sense because it is displayed after the
completion of the flow.

Change-Id: Iccc50867b28dd20ee45cad623ab19f23ff865682