Fix unplugging member ports
Unplugging a port from a VM can take some time before the update is
committed to the DB, 2 (quick) successive member deletions can trigger a
race condition when computing the delta of the ports that need to be
removed/added: Octavia may get a list of ports connected to an amphora
that is not correct.

This commit ensures that a port is also removed from the DB when
unplugging a member port.

Story 2009864
Task 44543

Change-Id: Iaa8ea5255b225ef87461c1f85b80f36210b1bde4