Fix potential race conditions on update requests in the v2 worker
The v2 worker was not waiting for a resource to be set to PENDING_UPDATE
before updating it effectively. In some circumstances (loaded
server/slow DB for instance), the resources may have been updated with
old data from the DB, resulting in a no-op update.

This wait for a PENDING_UPDATE state exists in amphorav1 but it was not
moved to v2 when we switched the logic from the DB-based objects to
dict-based objects.

Note for stable/yoga: this backport also contains [0] which improves
code coverage.

[0] I403e1861b1b51c07edad35c64ffac22167342c88

Story 2009887
Task 44645
Story 2009985
Task 45032

Change-Id: I433e074c2eac5f6875ee5dace8faf86969b24f66
(cherry picked from commit 6042de759db557f421482d22185c5e5075511417)