Prevent incorrect reschedule of resumed tasks with jobboard
A previous patch [0] introduced a keepalive mechanism that allows a
controller to notify taskflow that a task is being executed.
But it doesn't handle "resumed" tasks (tasks that have been rescheduled
from one controller to another controller).

This patch moves this keepalive task in a thread that is attached to a
taskflow conductor, the thread will extend the expiration of the tasks
that are executed by this conductor.

[0] I6619ba117f7051fe44086389789bc6e2810fd23d

Story 2009998
Task 45089

Change-Id: I29fcad9e121a30d6e8f8178f2f078cf10771a32a