Fix AttributeError in exception handler
The exception handler of save_state() caused an AttributeError
because it tried to get the output attribute of the Exception,
which does not exist in general. Added a test to cover this
scenario.

Change-Id: I36acdee1f8782b17c234f6b250facc5c8d0aaf87
(cherry picked from commit 384b83d88b436c8e657b09f97b12e42d7c9c339b)