Fix driver-agent cleanup
The cleanup function (that handles graceful shutdown) of the Octavia
driver-agent didn't work properly, the threads that were spawned for
stopping the Unix listeners were stuck (deadlock), so the threads leaked
and the join functions in the caller didn't return.

Now the code uses a thread for handling the connections, and the
shutdown function is called from the process.

Story 2010052
Task 45468

Change-Id: Ida33a1a24eab2706902b1ce3cca704d51fa6a599