Fix HealthMonitorToErrorOnRevertTask revert method
So far, whenever we had to revert HealthMonitorToErrorOnRevertTask,
we called mark_health_mon_prov_status_error with the wrong parameter
(health_mon[constants.POOL_ID] == the pool id).

This would set the pool's provisioning status to ERROR.

Later in the revert task we call mark_pool_prov_status_active with the
correct argument (health_mon[constants.POOL_ID] == the pool id), so we
set it back to ACTIVE, and the task continues correctly.

This means we might have missed setting the health monitor's
provisioning status to ERROR, and therefore we might have missed some
errors.

This commit changes the parameter to be the correct one in both API
v1 and v2, and it modifies the tests accordingly.

Story 2010113
Task 45690

Change-Id: I5a749c751d184ec10957a31b569419d57fe8daf6