Fix image creation for stable branches
A previous patch[1] renamed the udp-masquerade.sh file in the amphora-agent element to lvs-masquerade.sh as it is now also being used for the SCTP protocol in addition to UDP.
Unfortunately this broke cross-branch image building because the DIB install-static element will copy the file from the current checked out Octavia elements.
So if a user tries to build an image for stable/victoria using the "-g stable/victoria" flag, they will get a victoria amphora-agent, but the static files from the checked out version of Octavia.
This patch adds a symbolic link from the old file name to the new one to facilitate backwards compatibility with the older amphora-agents.

[1] https://review.opendev.org/c/openstack/octavia/+/753247

Story: 2010313
Task: 46379
Change-Id: I8b2862cf67d815309af3d13a68bab65e770775aa