Filter ComputeWaitTimeoutException when jobboard is disabled
A commit [0] has changed the behavior of the task that waits for a
compute server to be active. The task used to have a loop that was
waiting until the server was active, it is now using a Retry feature in
taskflow: the task throws an exception if the server is not active,
taskflow marks the task as failed, but it retries to execute it later
(until a max number of retries).
The exception thrown by the task should be hidden to the user, it is not
a real error, but it is part of the workflow. The exception is correctly
filtered out when jobboard is enabled, but it is not when jobboard is
disabled.

[0] I174e63cfe176e1792d0bbcd348dbc9115544e297

Story https://storyboard.openstack.org/#!/story/2010369
Task 46595

Change-Id: I282fa6a4ce43f0d2cb9cba48b7149ef66ede44fc
(cherry picked from commit 06f67f442b2a4f0c7e9d8578f8d964b613914bc4)