Align signature of AbstractNetworkDriver.update_vip method
The update_vip() method gets called in network_tasks.py without setting
the for_delete argument. That works because the actual network drivers
implement that function using a default for the for_delete argument.
This patch updates the signature of the respective base class in case a
driver does not overwrite this method.

Change-Id: I304e06d1d187070b260347926f3b516072827807