Add a newline when writing the server state file
Before restarting haproxy, the amphora-agent dumps the state of the
servers in a file, so when haproxy reloads, it can recover the operating
state of its backends' members.
But with haproxy 2.4 (in Centos 9 Stream), it fails with a "corrupted
global server state file" error when reading this file. It seems that
most recent haproxy releases have a more strict validation of the format
of the file and ensure that it is terminated with a newline.

This commit adds a newline to the file (all the responses to haproxy
queries are stripped in the amphora-agent). It fixes the issue on Centos
9 Stream (and turns off a warning on Ubuntu).

Story: 2010442
Task: 46873

Change-Id: I00e327e1d94e46aa13a38120df9865ec34eaa593