Ensure options are dumped in debug logs
This change makes the following processes dump loaded options when
the processes start.
 - octavia-driver-agent
 - octavia-health-manager
 - octavia-house-keeping

The actual options used by services are useful for debugging.

The loaded options are already logged by api and worker(*1), so this
ensures all daemons provide the consistent capability.

(*1)
octavia-worker depends on the cotyledon library, which internally
calls the (same) oslo.config functionality to dump options.

Change-Id: I79331dff469c6c81dd5afe04a950cdfd16abc3c1