Filter out details from taskflow logs with v2+jobboard
When enabling INFO-level logs in taskflow and using jobboard with
amphorav2, taskflow prints the string representation of a job when it is
completed. It includes the parameters of the flow, which might include
private information from TLS-enabled listeners and pools such as
certificates, private_key and intermediate certificates.

This commit filters out the private information from the logs by using
logging.Filter, it replaces private attributes with '***'.

Story 2010523
Task 47125

Change-Id: I2df8a49851feb1445b5128ce99b880ddb77782ad