allowed_cidr validation for additional_vips
The validation for the allowed_cidr parameter did not take into account
the IP version of additional vIPs. The parameter was rejected if the IP
version did not match the IP version of the primary vIP. As additional
vIPs can have a different IP version from the primary vIP all vIPs must
be checked during validation.

Change-Id: I3706fd5e12e17be37edce974563c6806d4f09709