Fix upgrade check not working
The octavia-status upgrade check command fails because
the initialization of Policy() requires oslo_config to
be initialized already which it isn't.

This can be reproduced in master by running:

  tox -e venv -- octavia-status upgrade check

I can observe this issue multiple relases back which
probably means this has been broken for a long time.

We shouldn't need to init Policy() because it just
loads policy rules which is not checked in upgrade
check anyway.

Change-Id: I16430fa52db02e7445203994220673c1037d764c