Fix test_driver_agent tests with newer octavia-lib
The tests in test_driver_agent compare recursively the dicts defined in
octavia with the dicts returned by octavia-lib. But when a new attribute
is added to octavia-lib, it breaks the tests in the -tips job until the
new attribute is included in octavia.

To mitigate this issue, the tests in octavia should only compare the
dicts by using the keys that are known by octavia.

Change-Id: I90ddebabcd90f30fccc6e1751f7c188b35ef16c4