Reduce duration of failovers with amphora in ERROR
In the failover flow, there are multiple tasks for the configuration of
VRRP for the other amphorae of the load balancer, but during outage the
other amps may not be available. To prevent the tasks from attempting
connections to unreachable amphorae, we can detect in the first task
that an amp is unreachable and pass this information to the other tasks.

Those connection attempts could have taken a lot of time, between 15 min
and 40 min depending on the configuration of Octavia and the provider
driver (amphorav1 or amphorav2)

Closes-Bug: #2033894

Note: stable/2023.1 and older, the patch also includes modifications in
      octavia/controller/worker/v1/

Conflicts:
	octavia/controller/worker/v2/tasks/amphora_driver_tasks.py

Change-Id: Ib33a0b8d2875e4ff97c65933fe9360bb06994d32
(cherry picked from commit ca70587f0bcb4b2a05bd5c243f111140a5762114)
(cherry picked from commit 6fdc2079da3392fdfbdf4a80df05bfe7d149ca95)