Fix incorrect masquerade rules in multivip LBs
The up scripts of the interface files were called only when the
interface moved from down to up, it means that they were not called
during the update of the configuration of an interface. So if during an
update, if an ipv6 subnet was plugged while the ipv4 subnet was already
there, the up script that sets the masquerade rules was not called.

It broke connectivity for either ipv4 or ipv6 between the client and the
members in UDP listeners in multivip load balancers.

Closes-Bug: #2037943

Change-Id: Iad78de1764bc3a3f699b5feef9e58999d2efe613
(cherry picked from commit 5c051a6ab6604c0cc38d59e1fdc0e6b199ed9932)
(cherry picked from commit 1f38d29b20f0769774517ae95d7d46c2eaff3265)