Fix race condition in cascade delete
update_vip was called multiple times concurrently when cascade deleting
a load balancer with many listeners, it may trigger a race condition
when fetching, computing and updating the SGs.
Calling update_vip for each listener is not necessary, it's now called
only once, that fixes the race condition and optimize the delete flow.

Closes-Bug: #2038529
Change-Id: I4a6e4830d0e916b1af1a11dd10097980a57d97ea
Signed-off-by: Gregory Thiemonge <gthiemon@redhat.com>