Disable seqdiag in the spec docs
The tls-data-security specs use seqdiag to describe communication between
some components.
seqdiag is unmaintained and doesn't work with recent Pillow releases (10.0.1).
Disable the generation of those diagrams as they are part of old specs.

Also remove the dependency on seqdiag

Disable octavia-grenade-skip-level which will be fixed in another commit

Partial-Bug: #2026345

Change-Id: I6b5c5d6c651dac223a205409c49085faf78cff15