Amphora multiple availability zones support
This change adds support to define multiple compute availability zones for the
amphora creation. Multiple ways to define a list of compute zones available:
* Define a list of Octavia availability zones in flavor.
* Define list of compute zones in octavia flavor
* define list of compute zones in nova section of the config file

Once we create a load balancer with multiple compute availability zones
(using either of the above ways), we calculate a zone for the initial master
amphora based on the loadbalancer_id hash. Then we pick another zone for the
failover amphora. Once we execute a failover, in the compute task we find
amphora which is going to be deleted using a status "pending delete" then put
a new amphora in exactly the same availability zone. This lets us preserve the
location of the amphoras.

Tests cover only basic cases with multiple compute zones. It might be needed
to add:
* Failover test that AZ stay the same
* Use octavia availability zones in flavor
* Use multiple conflicting definition (config plus flavor)

It might be better to do placement logic in controller_worker
(create_load_balancer and failover_amphora functions) instead of compute_tasks
and just define octavia availability_zone for the amphora during its creation,
but I might need help with this.