Unpin flake8-import-order
... to resolve the following ValueError.

```
ValueError: 'string' is not callable
```

The I202 test is disabled because multiple files need updates. It may
be enabled later once all the warnings are addressed.

Fixed E721 test errors which appears 4 times.

E721 do not compare types, for exact checks use `is` / `is not`, for
instance checks use `isinstance()`

Change-Id: I68880b83c8a555ed30c3ce18b73ae27b8db34dc8