Don't update the management port when calculating delta [v1]
CalculateDelta should be used to get the list of subnets/networks to
plug (or unplug) into the amphora for the VIP and the members.
It should never reconfigure the management port.

Story 2010623
Task 47564

This was originally fixed for *v2*, but not for v1 as this patch does.
Master v1 already removed from the code.

[1] https://review.opendev.org/c/openstack/octavia/+/876041

Change-Id: I43c1e1e48c86e2706fb8286ca11264434228e0b5