Fix port abandonment on failover flow revert
There was an issue in the failover flow, where if the failover flow reverted
after the VIP base (vrrp) port had been created, the revert would fail to
delete port as part of the revert. It did however log that it was not cleaning
up the port.
This patch corrects the issue such that VIP base ports get cleaned up on a
failover revert.

Change-Id: Iaf08c2bd229165638b4f7f850b26aaf899d540e1
(cherry picked from commit 0bfccba150da12555e847bcb0707573d99d00da8)