Update the Grafana dashboard for HAProxy 2.4+
HAProxy made some breaking changes to the metrics format for the
listener, pool, and memeber status metrics in version 2.4 and newer.
This patch updates the dashboard to use the new metric format and
updates the associated help text in the Prometheus exporter content.

This patch also corrects some typos for idle connections and SSL reuse.

Closes-Bug: #2065793
Change-Id: I6ba54e93f23961ede965dd1b4b3f58f73f7c1188