Fix amphora image builds to use DIB bindep
The amphora image builds are currently failing because we are trying to
install the binary packages for diskimage-builder ourselves in the
ansible for the job. This patch changes that to use the bindep.txt
included with diskimage-builder.

Change-Id: I8e489b251ff0e58809b66c0145023769abef6e68
(cherry picked from commit d682e74d33073dede4476e8de3e20766dc49063a)