Fix verification of certificates signed by a private CA
When using a certificate that was signed by a private CA (including
self-signed certs) on Neutron endpoints, certificate verification fails.

This was because the custom certificate was not used when creating
user's session even though the custom certificate file path is provided
in octavia.conf.

Fixing by feeding neutron cafile as a keyword argument.

Closes-Bug: #2046382
Co-authored-by: Ilia Kerbs <ikerbs@protonmail.com>

Change-Id: Ia92fa6140b0fc608281e846d0635dd28217f2630