Fix missing GARP with UDP listeners on SINGLE LB
GARP was not sent when using a SINGLE topology LB with a UDP listener.
Ensure that the GARP is sent when keepalived starts/reloads.

Note: as the keepalivedlvs configuration is populated only when a pool
is added, the GARP cannot be sent earlier (when the listener is added),
this behavior differs from the behavior of ACTIVE_STANDBY or
haproxy-based load balancers.

Closes-Bug: #2086195

Change-Id: I0ff05a27265b32d7e37b162a3e82108253e25530