Support rotating server_certs_key_passphrase key
The server_certs_key_passphrase was introduced to encrypt
the stored amphora certificates and the configuration option
for the passphrase has a default value.

This default value cannot be changed once assigned as there
is data stored encrypted using the passphrase, if the
passphrase is changed all amphora needs failover so that
the certificates is stored encrypted using the passphrase.

This wrap the existing cryptography.fernet.Fernet logic into
a utils helper function and converts it to using MultiFernet
so that the first key in the list is used for encrypting data
and all the others is available for decryption that means
that the the first key in the list can be changed and the
current key appended to the end of the list and when you
have organically (over time) performed amphora failovers you
can remove old keys.

The server_certs_key_passphrase config option is changed to
a ListOpt which is backward compatible with StrOpt.

We introduce a new FernetKeyOpt item_type for the config
option because ListOpt of strings (or if we would have
used a MultiStrOpt) does not support the `regex` kwarg
for regex based validation.

Closes-Bug: #2106810
Change-Id: Iadf100ab45499b59421425bc6874012a8bf0bde5